rfid_tracker/requirements.txt
# ========================================
# FILE: requirements.txt
# ========================================
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SocketIO==5.4.1
python-socketio==5.12.0
eventlet==0.40.3
pyserial==3.5
python-dotenv==1.2.0
requests==2.32.5
paramiko==3.4.0